% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{benford}
\alias{benford}
\title{Benford's law}
\usage{
benford(b)
}
\arguments{
\item{b}{Base, integer which has to be at least 2}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object describing Benford's law for the given base.
}
\examples{
# Distribution of first digits,according
# to Benford's law, in decimal system
benford(10)
}
\seealso{
Other distributions: 
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{bin}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
