% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{bernoulli}
\alias{bernoulli}
\title{The Bernoulli distribution}
\usage{
bernoulli(p)
}
\arguments{
\item{p}{Success probability}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the Bernoulli distribution with the given
parameters.
}
\examples{
# Distribution modelling a fair coin toss
bernoulli(1/2)
}
