% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{bin}
\alias{bin}
\title{The binomial distribution}
\source{
This function uses \code{\link[stats:Binomial]{stats::dbinom()}} to calculate the distribution
efficiently.
}
\usage{
bin(n, p)
}
\arguments{
\item{n}{Number of trials}

\item{p}{Success probability for each trial}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the binomial distribution with the given parameters.
}
\examples{
# How often a six can be expected when
# throwing a six-sided dice thrice
bin(3, 1/6)
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
