% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{cdf}
\alias{cdf}
\title{Cumulative distribution function}
\usage{
cdf(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A function.
}
\description{
Create an \code{R} function which behaves like the cumulative distribution
function of a given \code{ddf} distribution.
}
\details{
The cumulative distribution function \eqn{F_X} of a discrete random variable
\eqn{X} is given by
\deqn{F_X(x) = \mathbb{P}[X \le x],}
where \eqn{\mathbb{P}} is the probability operator.
}
\examples{
# Get value at 8 of the CDF of a geometric distribution
cdf(geometric(0.4))(8)
}
