% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{central_moment}
\alias{central_moment}
\title{Calculate central moments of distributions}
\usage{
central_moment(dist, n)
}
\arguments{
\item{dist}{\code{ddf} object, the distribution}

\item{n}{Order of the moment}
}
\value{
A double
}
\description{
This function calculates the \eqn{n}-th central moment, also known as the
\eqn{n}-th moment about the mean, of a \code{ddf} distribution.
}
\details{
The \eqn{n}-th central moment of a random variable \eqn{X} is given by
\deqn{E[(X - E[X])^n],}
where \eqn{E} is the expectation operator.
}
\examples{
# The zeroth central moment is always 1
central_moment(bin(5, 0.4), 0)

# The first central moment is always 0
central_moment(hypergeometric(10, 7, 5), n=1)

# The second central moment is the variance
central_moment(unif(10), 2)
# Result using the formula for the uniform distribution
(10^2-1)/12
}
\seealso{
Other moments: 
\code{\link{moment}()},
\code{\link{standardized_moment}()}
}
\concept{moments}
