% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve.R
\name{conv}
\alias{conv}
\title{Convolve two discrete distributions}
\usage{
conv(dist1, dist2, desc = "A convolution")
}
\arguments{
\item{dist1}{\code{ddf} object, the first distribution}

\item{dist2}{\code{ddf} object, the second distribution}

\item{desc}{The description for the resulting \code{ddf} object (optional)}
}
\value{
Convolution as a \code{ddf} object
}
\description{
Using this function, two distributions, given as \code{ddf} objects, can be
convolved.
}
\details{
Under the hood, this functions uses a \verb{C++} implementation for calculating
the convolution in order to get significant performance gains, compare
\code{\link[=convolve_cpp]{convolve_cpp()}}.
}
\examples{
# Calculate the distribution of the sum of throwing a dice twice
dice <- ddf(1:6)
conv(dice, dice, desc = "Distribution of throwing a dice twice")
}
\seealso{
\code{\link[=convolve_cpp]{convolve_cpp()}},\code{\link[=conv_n]{conv_n()}}
}
