% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve.R
\name{conv_n}
\alias{conv_n}
\title{Calculate n-fold convolution of a distribution with itself}
\usage{
conv_n(dist, n, desc = paste(n, "-fold convolution of", desc(dist)))
}
\arguments{
\item{dist}{A \code{ddf} distribution}

\item{n}{An integer specifying number of convolutions to perform}

\item{desc}{The description for the resulting \code{ddf} object (optional)}
}
\value{
n-fold convolution as a \code{ddf} object.
}
\description{
Using this function, a distribution, given as a \code{ddf} object, can be
convolved with itself n times.
}
\details{
For more details, see \code{\link[=conv]{conv()}}.
}
\examples{
# Calculate the distribution of the sum of throwing a dice four times
conv_n(ddf(1:6), 4, desc = "Distribution of throwing a dice four times")
}
\seealso{
Other {convolution functions}: 
\code{\link{conv}()},
\code{\link{convolve_cpp}()}
}
\concept{{convolution functions}}
