% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolve_cpp}
\alias{convolve_cpp}
\title{Convolve two discrete distributions}
\usage{
convolve_cpp(supp1, probs1, supp2, probs2)
}
\arguments{
\item{supp1, probs1}{Numeric vectors.}

\item{supp2, probs2}{Numeric vectors.}
}
\value{
A list.
}
\description{
\code{convolve_cpp} returns the convolution of two discrete distributions given
their supports and the corresponding probabilities.
}
\details{
Some assumptions are made for the function to work properly which should be
followed. Most importantly, corresponding numeric vectors (e.g. \code{supp1} and
\code{probs1}) should have the same length.

The function returns a list with components "support" and "probabilities"
which give the matching attributes of the convolution.
}
\examples{
# Calculate the probability distribution of the sum of two six-sided dice
convolve_cpp((1:6), rep(1/6, 6), (1:6), rep(1/6, 6))
}
