% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-class.R
\docType{class}
\name{ddf-class}
\alias{ddf-class}
\title{An S4 class to represent a discrete distribution with finite support}
\description{
A discrete distribution with finite support is here represented by its
probability mass function (PMF). For this, it has one slot which stores the
support of the PMF as a numeric vector and another slot for storing the value
of the PMF at the corresponding point in the support.
There is one further slot for a description.
}
\details{
New \code{ddf} objects are checked for validity to ensure that they fulfill basic
properties of a discrete probability distribution with finite support.
The relevant checks for this are:
\itemize{
\item Support and probability vectors must have the same length.
\item All probabilities have to lie between 0 (excluded) and 1 (included).
Here, 0 is not included as elements of the support have, by definition,
non-zero probability.
\item The probabilities have to sum up to approximately 1. Currently the
tolerance for this is set to \code{1e-10}.
}
}
\section{Slots}{

\describe{
\item{\code{supp}}{A numeric vector, the support.}

\item{\code{probs}}{A numeric vector, the corresponding probabilities.}

\item{\code{desc}}{A character, a short description of the discrete distribution.}
}}

