% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-class.R
\name{-,ddf,missing-method}
\alias{-,ddf,missing-method}
\title{Multiply support of a distribution with -1}
\usage{
\S4method{-}{ddf,missing}(e1, e2)
}
\arguments{
\item{e1}{A \code{ddf} distribution}

\item{e2}{Unused argument. Has to be missing.}
}
\value{
A \code{ddf} distribution.
}
\description{
Multiply support of a distribution with -1
}
\examples{
-bin(5, 0.1)
}
