% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{decile}
\alias{decile}
\title{Calculate deciles of distributions}
\usage{
decile(dist, k, method = "all")
}
\arguments{
\item{dist}{\code{ddf} object, the distribution}

\item{k}{Integer between 0 and 10, exclusive}

\item{method}{Character specyfing output format}
}
\value{
A numeric vector
}
\description{
This function calculates the \eqn{k}-th decile of a \code{ddf} distribution.
}
\details{
For details consult the details of \code{\link[=quantile]{quantile()}} and recall that a \eqn{k}-th
decile is a \eqn{k/10}-quantile and vice versa.
}
\examples{
decile(bin(10, 0.3), 3)
}
\seealso{
Other quantiles: 
\code{\link{percentile}()},
\code{\link{quantile}()},
\code{\link{quartile}()}
}
\concept{quantiles}
