% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-class.R
\name{desc<-}
\alias{desc<-}
\alias{desc<-,ddf-method}
\title{Set the description of a \code{ddf} object}
\usage{
desc(dist) <- value

\S4method{desc}{ddf}(dist) <- value
}
\arguments{
\item{dist}{A \code{ddf} object}

\item{value}{A character, new description}
}
\value{
Description text of the \code{ddf} object.
}
\description{
Set the description of a \code{ddf} object
}
\examples{
mydist <- unif(8)
print(mydist)
desc(mydist) <- "This distribution needs a better description"
print(mydist)
}
