% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{distribution_range}
\alias{distribution_range}
\title{Calculate the range of distributions}
\usage{
distribution_range(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A double.
}
\description{
This function calculates the range of a \code{ddf} distribution.
}
\details{
The range is the length of the narrowest interval containing the entire
support.
}
\examples{
# Range of the Rademacher distribution
distribution_range(rademacher())
}
