% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-special-moments.R
\name{excess_kurtosis}
\alias{excess_kurtosis}
\title{Calculate the excess kurtosis of distributions}
\usage{
excess_kurtosis(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A double.
}
\description{
This function calculates the excess kurtosis of a \code{ddf} distribution.
}
\details{
The excess kurtosis is simply defined as the \code{\link[=kurtosis]{kurtosis()}} minus 3, hence we
only refer to the linked help page for some more details.
}
\examples{
excess_kurtosis(unif(8))
}
\seealso{
\code{\link[=kurtosis]{kurtosis()}}
}
