% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central-tendency.R, R/ddf-class.R
\name{expected_value}
\alias{expected_value}
\alias{mean,ddf-method}
\title{Calculate the expected value of distributions}
\usage{
expected_value(dist)

\S4method{mean}{ddf}(x)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}

\item{x}{\code{ddf} object, the distribution.}
}
\value{
A double.
}
\description{
This function calculates the mean/expected value of a discrete distribution
with finite support given as a \code{ddf} object.
}
\details{
The expected value of a discrete random variable \eqn{X} with finite support
\eqn{\{x_1, \dots, x_n\}} and corresponding probabilities
\eqn{p_1, \dots, p_n} is defined as the weighted average of the \eqn{x_k}
values with weights \eqn{p_k},
\deqn{E[X] := \sum_{k=1}^n x_k p_k.}
As we only work with discrete distributions with finite support, we don't
provide a more general definition using the tools from measure theory and
instead content ourselves with the basic one above.

Note that there are two ways to use this function, one being
\code{expected_value()} and the other one being the S4 generic method \code{mean()}.
See also ‘Examples.’
}
\examples{
# Expected value when throwing a six-sided dice
expected_value(unif(6))
# Equivalent result using generic method
mean(unif(6))
}
\seealso{
\code{\link[=mean]{mean()}}
}
