% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{hypergeometric}
\alias{hypergeometric}
\title{The hypergeometric distribution}
\source{
This function uses \code{\link[stats:Hypergeometric]{stats::dhyper()}} to calculate the distribution
efficiently.
}
\usage{
hypergeometric(N, K, n)
}
\arguments{
\item{N}{A positive integer, the population size.}

\item{K}{A non-negative integer, the number of success states in the
population. Has to be less or equal to \code{N}.}

\item{n}{A non-negative integer, the number of draws. Has to be less or equal
to \code{N}.}
}
\value{
A \code{ddf} distribution as described above.
}
\description{
Create a \code{ddf} object for the hypergeometric distribution with the given
parameters.
}
\details{
The hypergeometric distribution models the number of successes when drawing
\eqn{n} elements, \emph{without} replacement, from a finite population of size
\eqn{N} which contains \eqn{K} success states.

It has support
\deqn{\{\max(0, n+K-N), \dots, \min(n, K)\}}
on which its probability mass function is given by
\deqn{p(k) = \frac{\binom{K}{k} \binom{N-K}{n-k}}{\binom{N}{n}}.}
}
\examples{
# Model the total number of blue marbles when
# drawing 5 marbles from an urn containing
# 6 blue marbles and 20 marbles in total
hypergeometric(20, 6, 5)
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{bin}()},
\code{\link{geometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
