% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{iqr}
\alias{iqr}
\title{Calculate the interquartile range of distributions}
\usage{
iqr(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A double.
}
\description{
This function calculates the interquartile range of a \code{ddf} distribution.
}
\details{
The interquartile range is defined as the difference \eqn{Q_3 - Q_1}, where
\eqn{Q_1} and \eqn{Q_3} are the first and third quartiles, respectively (for
more information see \code{\link[=quartile]{quartile()}} and \code{\link[=quantile]{quantile()}}).
}
\examples{
# Interquartile range of a uniform distribution
iqr(unif(10))
}
