% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-special-moments.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate the kurtosis of distributions}
\usage{
kurtosis(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A double.
}
\description{
This function calculates the kurtosis of a \code{ddf} distribution.
}
\details{
The kurtosis is given by the fourth standardized moment, thus for some more
details one can consult \code{\link[=standardized_moment]{standardized_moment()}}.

It measures the "tailedness" of a distribution, i.e. the heaviness
of its tails. A higher kurtosis indicates a higher propensity to produce
outliers.
}
\examples{
# The rademacher distribution has a very low kurtosis
# (in fact it is as minimal as possible)
kurtosis(rademacher())
# A binomial distribution with relatively low probability
# has rather high kurtosis
kurtosis(bin(10, 1e-4))
}
