% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_tendency.R
\name{median}
\alias{median}
\title{Calculate medians of distributions}
\usage{
median(dist, method = "all")
}
\arguments{
\item{dist}{\code{ddf} object, the distribution}

\item{method}{Character specyfing output format}
}
\value{
A numeric vector
}
\description{
This function calculates the median of a distribution, given as a \code{ddf} object.
In the case the median is not unique, by default all possible values are
returned. This behaviour can be controlled using the \code{methods} argument.
}
\details{
For details consult the details of \code{\link[=quantile]{quantile()}} and recall that a median is a
\eqn{\frac{1}{2}}-quantile and vice versa.
}
\examples{
# In case the mode is not unique, by default all
# possible modes are returned
median(unif(4))
# This can be controlled using the method argument
median(unif(4), method = "mean")
}
