% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central-tendency.R
\name{medians}
\alias{medians}
\title{Calculate medians of distributions}
\usage{
medians(dist, method = "all")
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}

\item{method}{A character specifying output format.}
}
\value{
A numeric vector.
}
\description{
This function calculates the median(s) of a distribution, given as a \code{ddf}
object. In the case of the median not being unique, by default all possible
values are returned. This behaviour can be controlled using the \code{methods}
argument.
}
\details{
For details consult the details of \code{\link[=quantile]{quantile()}} and recall that a median is a
\eqn{\frac{1}{2}}-quantile and vice versa.
}
\examples{
# In case the mode is not unique, by default all
# possible modes are returned
medians(unif(4))
# This can be controlled using the method argument
medians(unif(4), method = "mean")
}
