% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_tendency.R
\name{mode}
\alias{mode}
\title{Calculate mode(s) of distributions}
\usage{
mode(dist, method = "all")
}
\arguments{
\item{dist}{\code{ddf} object, the distribution}

\item{method}{Character specyfing output format}
}
\value{
A numeric vector
}
\description{
This function calculates the mode of a distribution, given as a \code{ddf} object.
In the case the mode is not unique, by default all possible values are
returned. This behaviour can be controlled using the \code{methods} argument.
}
\details{
A mode of a discrete random variable \eqn{X} is a value at which the
probability mass function of \eqn{X} takes its maximal value.

In general, it is not unique which is the reason for this function to have
the \code{method} argument.
\itemize{
\item If \code{method} is set to "all", all values in the support which are a mode of
the given distribution are returned. This is the default behaviour.
\item If \code{method} is set to "mean", the mean of all such values is returned.
\item If \code{method} is set to "min" or "max", the respective elements of all
modes are returned.
}
}
\examples{
# In case the mode is not unique, by default all
# possible modes are returned
mode(unif(3))
# This can be controlled using the method argument
mode(unif(3), method = "mean")
}
