% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moment}
\alias{moment}
\title{Calculate moments of distributions}
\usage{
moment(dist, n)
}
\arguments{
\item{dist}{\code{ddf} object, the distribution}

\item{n}{Order of the moment}
}
\value{
A double
}
\description{
This function calculates the \eqn{n}-th (raw) moment, also known as the
\eqn{n}-th moment about zero, of a \code{ddf} distribution.
}
\details{
The \eqn{n}-th (raw) moment of a random variable \eqn{X} is given by
\deqn{E[X^n],}
where \eqn{E} is the expectation operator.
}
\examples{
# Expected value when throwing a six-sided dice
moment(unif(6), 1)
}
\seealso{
Other moments: 
\code{\link{central_moment}()},
\code{\link{standardized_moment}()}
}
\concept{moments}
