% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{percentile}
\alias{percentile}
\title{Calculate percentiles of distributions}
\usage{
percentile(dist, k, method = "all")
}
\arguments{
\item{dist}{\code{ddf} object, the distribution}

\item{k}{Integer between 0 and 100, exclusive}

\item{method}{Character specyfing output format}
}
\value{
A numeric vector
}
\description{
This function calculates the \eqn{k}-th percentile of a \code{ddf} distribution.
}
\details{
For details consult the details of \code{\link[=quantiles]{quantiles()}} and recall that a \eqn{k}-th
percentile is a \eqn{k/100}-quantile and vice versa.
}
\examples{
percentile(bin(10, 0.3), 30)
}
\seealso{
Other quantiles: 
\code{\link{decile}()},
\code{\link{quantiles}()},
\code{\link{quartile}()}
}
\concept{quantiles}
