% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-class.R
\name{plot,ddf,missing-method}
\alias{plot,ddf,missing-method}
\title{Plot a distribution}
\usage{
\S4method{plot}{ddf,missing}(
  x,
  xlab = "support",
  ylab = "probabilities",
  col = "deepskyblue3",
  main = NULL,
  sub = NULL,
  y
)
}
\arguments{
\item{x}{A \code{ddf} object, the distribution.}

\item{xlab}{The label for the \eqn{x}-axis.}

\item{ylab}{The label for the \eqn{y}-axis.}

\item{col}{The color for the points of the plot.}

\item{main}{An overall title of the plot.}

\item{sub}{A subtitle for the plot.}

\item{y}{Unused argument. Has to be missing.}
}
\value{
A \code{ggplot} object displaying the given distribution.
}
\description{
Create a plot of a \code{ddf} distribution.
}
\examples{
plot(pois(8))
}
\seealso{
Other plotting functions: 
\code{\link{plot_cdf}()},
\code{\link{plot_pmf}()}
}
\concept{plotting functions}
