% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cdf}
\alias{plot_cdf}
\title{Plot the CDF of a distribution}
\usage{
plot_cdf(
  dist,
  xlab = "x",
  ylab = "P[X ≤ x]",
  col = "deepskyblue3",
  main = NULL,
  sub = NULL
)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}

\item{xlab}{The label for the \eqn{x}-axis.}

\item{ylab}{The label for the \eqn{y}-axis.}

\item{col}{The color for the points of the plot.}

\item{main}{An overall title of the plot.}

\item{sub}{A subtitle for the plot.}
}
\value{
A \code{ggplot} object displaying the CDF of the given distribution.
}
\description{
This function creates a plot of the cumulative distribution function (CDF) of
a \code{ddf} distribution.
}
\details{
For further details on the cumulative density function, you may consult \code{\link[=cdf]{cdf()}}.
}
\examples{
# Plot the CDF of a binomial distribution
plot_cdf(bin(20, 0.8))
}
\seealso{
Other plotting functions: 
\code{\link{plot,ddf,missing-method}},
\code{\link{plot_pmf}()}
}
\concept{plotting functions}
