% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pmf}
\alias{plot_pmf}
\title{Plot the PMF of a distribution}
\usage{
plot_pmf(
  dist,
  xlab = "x",
  ylab = "P[X = x]",
  col = "deepskyblue3",
  main = NULL,
  sub = NULL
)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}

\item{xlab}{The label for the \eqn{x}-axis.}

\item{ylab}{The label for the \eqn{y}-axis.}

\item{col}{The color for the points of the plot.}

\item{main}{An overall title of the plot.}

\item{sub}{A subtitle for the plot.}
}
\value{
A \code{ggplot} object displaying the PMF of the given distribution.
}
\description{
This function creates a plot of the probability mass function of a \code{ddf}
distribution.
}
\details{
For further details on the probability mass function, you may consult \code{\link[=pmf]{pmf()}}.

You might also be interested in the generic \code{\link[base:plot]{base::plot()}} method for \code{ddf}
objects.
}
\examples{
# Plot the PMF of a binomial distribution
plot_pmf(bin(20, 0.8))
# A similar plot using the S4 generic
plot(bin(20, 0.8))
}
\seealso{
Other plotting functions: 
\code{\link{plot,ddf,missing-method}},
\code{\link{plot_cdf}()}
}
\concept{plotting functions}
