% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{pmf}
\alias{pmf}
\title{Probability mass function}
\usage{
pmf(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A function.
}
\description{
Create an \code{R} function which behaves like the probability mass function of a
given \code{ddf} distribution.
}
\details{
The probability mass function \eqn{p_X} of a discrete random variable \eqn{X}
is given by
\deqn{p_X(x) = \mathbb{P}[X = x],}
where \eqn{\mathbb{P}} is the probability operator.
}
\examples{
# Get probability of 5 in a geometric distribution
pmf(geometric(0.4))(5)
}
