% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-approximated.R
\name{pois}
\alias{pois}
\title{The Poisson distribution}
\source{
In order to calculate a fitting cutoff, the quantile function
\code{\link[stats:Poisson]{stats::qpois()}} is used. \code{\link[stats:Poisson]{stats::dpois()}} is then employed to calculate
the distribution.
}
\usage{
pois(lambda, eps = 1e-10)
}
\arguments{
\item{lambda}{Expected rate of occurrences}

\item{eps}{How close the distribution is approximated (see details, optional)}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the Poisson distribution with the given parameters.
}
\details{
As the Poisson distribution has countably infinite support and this package
only works with discrete distributions with finite support, the resulting
\code{ddf} object approximates the Poisson distribution.

For this, the support is cut off at a large enough integer such that the
overall probability is still close to 1.
The cutoff is controlled via the \code{eps} argument which specifies how close the
sum of all probabilities has to be to 1. The default value is 1e-10 since
this is also the minimum accuracy required for creating valid \code{ddf} objects.
}
\examples{
pois(0.3)
# A more accurate approximation of the same distribution
pois(0.3, 1e-15)
}
