% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{quantile}
\alias{quantile}
\title{Calculate quantiles of distributions}
\usage{
quantile(dist, p, method = "all")
}
\arguments{
\item{dist}{\code{ddf} object, the distribution}

\item{p}{Number between 0 and 1, exclusive}

\item{method}{Character specyfing output format}
}
\value{
A numeric vector
}
\description{
This function calculates \eqn{p}-quantiles of a \code{ddf} distribution.
}
\details{
A \eqn{p}-quantile of a random variable \eqn{X} is a value \eqn{x} such
that
\deqn{\mathbb{P}[ X \le x] \ge p \qquad\text{ and }\qquad \mathbb{P}[X \ge x] \ge 1-p,}
where \eqn{\mathbb{P}} denotes the probability operator.

In general, it is not unique which is the reason for this function to have
the \code{method} argument.
\itemize{
\item If \code{method} is set to "all", all values in the support which are a \eqn{p}-
quantile of the given distribution are returned. This is the default behaviour.
\item If \code{method} is set to "mean", the mean of all such values is returned.
\item If \code{method} is set to "min" or "max", the respective elements of the
\eqn{p}-quantiles in the support are returned.
}
}
\examples{
quantile(unif(6), 1/2)
# Using another method for output
quantile(unif(6), 1/2, method = "mean")
}
\seealso{
Other quantiles: 
\code{\link{decile}()},
\code{\link{percentile}()},
\code{\link{quartile}()}
}
\concept{quantiles}
