% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{quartile}
\alias{quartile}
\title{Calculate quartiles of distributions}
\usage{
quartile(dist, k, method = "all")
}
\arguments{
\item{dist}{\code{ddf} object, the distribution}

\item{k}{Integer between 0 and 4, exclusive}

\item{method}{Character specyfing output format}
}
\value{
A numeric vector
}
\description{
This function calculates the \eqn{k}-th quartile of a \code{ddf} distribution.
}
\details{
For details consult the details of \code{\link[=quantile]{quantile()}} and recall that a \eqn{k}-th
quartile is a \eqn{k/4}-quantile and vice versa.
}
\examples{
quartile(bin(10, 0.3), 3)
}
\seealso{
Other quantiles: 
\code{\link{decile}()},
\code{\link{percentile}()},
\code{\link{quantile}()}
}
\concept{quantiles}
