% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rademacher}
\alias{rademacher}
\title{The Rademacher distribution}
\usage{
rademacher()
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the Rademacher distribution with the given
parameters.
}
\examples{
rademacher()
}
