% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generate a report for a distribution}
\usage{
report(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A character.
}
\description{
This function writes a detailed report given a \code{ddf} distribution involving
most properties such as various measures of central tendency and dispersion,
skewness and kurtosis and more.
}
\examples{
# Write a report on a binomial distribution
# and output it using cat
cat(report(bin(20, 0.4)))
}
