% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples}
\alias{samples}
\title{Random samples from a distribution}
\source{
To generate the samples, \code{\link[base:sample]{base::sample()}} is used.
}
\usage{
samples(dist, n)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}

\item{n}{A non-negative integer giving the number of items to choose.}
}
\value{
A vector of length \code{size} with elements drawn from the distributions
support.
}
\description{
This function generates random samples of a given size based on a \code{ddf}
distribution.
}
\examples{
# Simulate throwing a dice 5 times
samples(unif(6), 5)
}
