% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{shift}
\alias{shift}
\title{Shift a distribution}
\usage{
shift(dist, s, desc = NULL)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}

\item{s}{A number, the shift to be applied to the distribution.}

\item{desc}{A character, the description for the shifted distribution.
(optional)}
}
\value{
A \code{ddf} object, the shifted distribution.
}
\description{
Shift (the support of) a \code{ddf} distribution by the specified amount.
}
\examples{
# Distribution for a simple symmetric random walk with
# step size 1 of length 4, centered at 3 (instead of 0)
shift(conv_n(rademacher(), 4), 3)
}
