% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-special-moments.R
\name{skew}
\alias{skew}
\title{Calculate the skewness of distributions}
\usage{
skew(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A double.
}
\description{
This function calculates the skewness of a \code{ddf} distribution.
}
\details{
"Skewness" here refers to \strong{Fisher's moment coefficient} of skewness, also
known under \strong{Pearson's} instead of Fisher's name or simply as the
\strong{moment coefficient of skewness}. It is given by the third standardized
moment, see also the help \code{\link[=standardized_moment]{standardized_moment()}}.

A distribution with negative skew has a longer left tail, its mass is
concentrated to the right end of its support and when plotting its
probability mass function one gets a right-leaning curve.

Analogously, a distribution with positive skew has a longer right tail, its
mass is concentrated to the left end of its support and when plotting its
probability mass function one gets a left-leaning curve.
}
\examples{
# A binomial distribution with success probability
# p < 0.5 has positive skew
skew(bin(10, 0.25))
# whereas one with p > 0.5 has negative skew
skew(bin(10, 0.75))
}
