% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Calculate the standard deviation of distributions}
\usage{
standard_deviation(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A double.
}
\description{
This function calculates the standard deviation of a \code{ddf} distribution.
}
\details{
The standard deviation is the square root of the variance (which itself is
the second central moment). For more details see \code{\link[=variance]{variance()}} or better directly
\code{\link[=central_moment]{central_moment()}}.
}
\examples{
# Standard deviation of a hypergeometric distribution
standard_deviation(hypergeometric(20, 7, 3))
}
