% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{standardized_moment}
\alias{standardized_moment}
\title{Calculate standardized moments of distributions}
\usage{
standardized_moment(dist, n)
}
\arguments{
\item{dist}{\code{ddf} object, the distribution}

\item{n}{Order of the moment}
}
\value{
A double
}
\description{
This function calculates the \eqn{n}-th standardized moment of a \code{ddf}
distribution.
}
\details{
The \eqn{n}-th standardized moment of a random variable \eqn{X} is given by
\deqn{\mu_n/\sigma^n,}
where \eqn{\mu_n} denotes the \eqn{n}-th moment about the mean (see also
\code{\link[=central_moment]{central_moment()}}) and \eqn{\sigma^n} is the \eqn{n}-th power of the standard
deviation.

Note that we here normalize with \eqn{\sigma^n}, although sometimes other
normalizations are used.
}
\examples{
# The second standardized moment is always 1
standardized_moment(unif(10), 2)

# The third standardized moment measures skewness
standardized_moment(bin(10, 0.8), 3)
}
\seealso{
Other moments: 
\code{\link{central_moment}()},
\code{\link{moment}()}
}
\concept{moments}
