% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ddf.R
\name{supp}
\alias{supp}
\alias{supp,ddf-method}
\title{Get the support of a \code{ddf} object}
\usage{
supp(dist)

\S4method{supp}{ddf}(dist)
}
\arguments{
\item{dist}{A \code{ddf} object}
}
\value{
Support as a numerical vector.
}
\description{
Get the support of a \code{ddf} object
}
\examples{
supp(ddf((1:3)))
}
