% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{unif}
\alias{unif}
\title{The discrete uniform distribution}
\usage{
unif(n, start_at_one = TRUE)
}
\arguments{
\item{n}{An integer, the upper end of the support.}

\item{start_at_one}{Logical, whether to start the support at 0 or 1.
(default: TRUE)}
}
\value{
A \code{ddf} distribution as described above.
}
\description{
Create a \code{ddf} object for the discrete uniform distribution on the first
\eqn{n} natural numbers (including or excluding 0).
}
\details{
The discrete uniform distribution describes models in which a finite number
of possible outcomes are equally likely to happen.
In general, it has an arbitrary finite set, say of cardinality
\eqn{N\in\mathbb{N}}, as its support with every element of the support having
the same probability \eqn{\frac{1}{N}}.

This function, however, only generates the uniform distribution on support
\eqn{\{1, \dots, n\}} or \eqn{\{0, \dots, n\}}. For more complicated supports,
it's expected to be much easier to simply use \code{\link[=ddf]{ddf()}} without specifying
probabilities, rather than using a custom function like this one.
For a demonstration of this, see also the ‘Examples’ below.
}
\examples{
# Probability distribution for modelling a
# six-sided dice (uniform on {1, ..., 6})
unif(6)

# For more complicated supports, use ddf
# Same distribution as above (except for description):
ddf(1:6)
# Not possible using unif:
ddf(seq(3, 12, 3))
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{bin}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{zipf}()}
}
\concept{distributions}
