% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{variance}
\alias{variance}
\title{Calculate the variance of distributions}
\usage{
variance(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A double.
}
\description{
This function calculates the variance of a \code{ddf} distribution.
}
\details{
The variance is the second central moment. For more details see the
corresponding help, i.e. \code{\link[=central_moment]{central_moment()}}.

It is a measure of dispersions, i.e. it measures how much the mass of a
distribution is spread around its mean. A lower variance indicates that
values are to be expected closer to the mean.
}
\examples{
# Variance of a binomial distribution
variance(bin(10, 0.3))
# Using the corresponding formula
10 * 0.3 * (1 - 0.3)
}
