% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{zipf}
\alias{zipf}
\title{The Zipf distribution}
\usage{
zipf(N, s = 1)
}
\arguments{
\item{N}{Total number of elements}

\item{s}{Exponent for the inverse power law (see details, default: 1)}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the Zipf distribution with the given parameters.
}
\examples{
zipf(5)
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{bin}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()}
}
\concept{distributions}
