% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{benford}
\alias{benford}
\title{Benford's law}
\usage{
benford(b)
}
\arguments{
\item{b}{An integer, the base of the number system. Has to be at least 2.}
}
\value{
A \code{ddf} distribution as described above.
}
\description{
Create a \code{ddf} object describing Benford's law for the given base.
}
\details{
Benford's law describes the phenomenon that in many real-life numerical data
sets, leading digits tend to be small.

For this, given an arbitrary base \eqn{b}, one can consider the distribution
with support \eqn{\{1, \dots, b-1\}} on which the probability mass function
\deqn{p(k) = \log_b \left( 1 + \frac{1}{k} \right)}
is defined, where the subscript \eqn{b} denotes the base of the logarithm.

For more information, we refer to the below linked Wikipedia article.
}
\examples{
# Expected distribution of first digits,according
# to Benford's law, in the decimal system
benford(10)
}
\references{
\url{https://en.wikipedia.org/wiki/Benford\%27s_law}
}
\seealso{
Other distributions: 
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{bin}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
