% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{bernoulli}
\alias{bernoulli}
\title{The Bernoulli distribution}
\usage{
bernoulli(p)
}
\arguments{
\item{p}{Success probability}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the Bernoulli distribution with the given
parameters.
}
\examples{
# Distribution modelling a fair coin toss
bernoulli(1/2)
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{beta_binomial}()},
\code{\link{bin}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
