% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{beta_binomial}
\alias{beta_binomial}
\title{The beta-binomial distribution}
\usage{
beta_binomial(n, alpha, beta)
}
\arguments{
\item{n}{Number of trials}

\item{alpha, beta}{Shape parameters for the underlying beta distribution}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the beta-binomial distribution with the given
parameters.
}
\examples{
beta_binomial(4, 0.6, 2)
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{bernoulli}()},
\code{\link{bin}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
