% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{beta_binomial}
\alias{beta_binomial}
\title{The beta-binomial distribution}
\usage{
beta_binomial(n, alpha, beta)
}
\arguments{
\item{n}{Number of trials}

\item{alpha, beta}{Shape parameters for the underlying beta distribution}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the beta-binomial distribution with the given
parameters.
}
\examples{
beta_binomial(4, 0.6, 2)
}
