% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{bin}
\alias{bin}
\title{The binomial distribution}
\source{
This function uses \code{\link[stats:Binomial]{stats::dbinom()}} to calculate the distribution
efficiently.
}
\usage{
bin(n, p)
}
\arguments{
\item{n}{An integer, the number of trials.}

\item{p}{A number between 0 and 1, the success probability for each trial.}
}
\value{
A \code{ddf} distribution as described above.
}
\description{
Create a \code{ddf} object for the binomial distribution with the given parameters.
}
\details{
The binomial distribution with parameters \eqn{n} and \eqn{p} models the
number of successes when conducting \eqn{n} independent experiments with each
one being distributed according to the Bernoulli distribution with success
probability \eqn{p} (see also \code{\link[=bernoulli]{bernoulli()}}).

It has support \eqn{\{0, \dots, n\}} on which its probability mass function
is given by
\deqn{p(k) = \binom{n}{k} p^k (1-p)^{n-k}.}
}
\examples{
# How often a six can be expected when
# throwing a six-sided dice thrice
bin(3, 1/6)
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
