% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve.R
\name{conv_n}
\alias{conv_n}
\title{Calculate the n-fold convolution of a distribution with itself}
\usage{
conv_n(dist, n, desc = NULL)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}

\item{n}{An integer specifying the number of convolutions to perform.}

\item{desc}{The description for the resulting \code{ddf} object. (optional)}
}
\value{
\eqn{n}-fold convolution as a \code{ddf} object.
}
\description{
Using this function, a distribution, given as a \code{ddf} object, can be
convolved with itself \eqn{n} times.
}
\details{
For more details, see \code{\link[=conv]{conv()}}.
}
\examples{
# Calculate the distribution of the sum of
# throwing a dice four times
conv_n(ddf(1:6), 4, desc = "Distribution of throwing a dice four times")
# Model a symmetric random walk starting at 0
# with step size 1 of length 4
conv_n(rademacher(), 4)
}
\seealso{
Other {convolution functions}: 
\code{\link{conv}()},
\code{\link{convolve_cpp}()}
}
\concept{{convolution functions}}
