% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-class.R
\name{ddf}
\alias{ddf}
\title{Create new \code{ddf} objects}
\usage{
ddf(
  supp,
  probs = rep(1/length(supp), length(supp)),
  desc = "A discrete distribution with finite support"
)
}
\arguments{
\item{supp}{A numeric vector representing the support}

\item{probs}{A numeric vector containing the probabilities
corresponding to the elements of the support}

\item{desc}{A short description of the discrete distribution}
}
\value{
A \code{ddf} object with the specified properties
}
\description{
Use this function to create a new \code{ddf} object by passing the support and
the corresponding probabilities as vectors to it.
}
\details{
This function makes sure that your \code{ddf} objects are "clean" by ensuring the
following two conditions:
\itemize{
\item The passed support is ordered in ascending order.
\item Duplicates within the support are removed and the probabilities of the
affected elements are simply summed up.
}

If the probabilities aren't supplied, they are distributed uniformly over the
specified support.
}
\examples{
# Create ddf object for an ordinary six-sided dice
dice <- ddf(1:6, desc = "Distribution modelling a six-sided dice")
# Create ddf object for an unfair coin toss without custom description
coin_toss <- ddf(c(1,2), c(1/4, 3/4))
}
