% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddfr-package.R
\docType{package}
\name{ddfr-package}
\alias{ddfr}
\alias{ddfr-package}
\title{ddfr: Work with discrete distributions with finite support in R}
\description{
This package enables you to create objects for representing discrete distributions with finite support by specifying their probability mass functions. These objects can then be used for a variety of tasks, from simple statistical analyses over plotting to convolutions.
}
\author{
\strong{Maintainer}: fdf \email{171835561+fdf-uni@users.noreply.github.com}

}
\keyword{internal}
