% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-class.R
\name{desc}
\alias{desc}
\alias{desc,ddf-method}
\title{Get the description of a \code{ddf} object}
\usage{
desc(dist)

\S4method{desc}{ddf}(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
A character, the description text of the \code{ddf} object.
}
\description{
This function returns the description of a given distribution.
}
\examples{
desc(ddf(1:3))
desc(ddf(1:3, desc = "My custom description"))
}
