% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central-tendency.R
\name{expected_value}
\alias{expected_value}
\title{Calculate expected values of distributions}
\usage{
expected_value(dist)
}
\arguments{
\item{dist}{\code{ddf} object, the distribution}
}
\value{
A double
}
\description{
This function calculates the mean/expected value of a discrete distribution
with finite support given as a \code{ddf} object.
}
\details{
Alternatively, you can also simply use the generic method \code{mean()}.
}
\examples{
# Expected value when throwing a six-sided dice
expected_value(unif(6))
# Equivalent result using generic method
mean(unif(6))
}
\seealso{
\code{\link[=mean]{mean()}}
}
