% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-approximated.R
\name{geometric}
\alias{geometric}
\title{The geometric distribution}
\source{
In order to calculate a fitting cutoff, the quantile function
\code{\link[stats:NegBinomial]{stats::qnbinom()}} is used. \code{\link[stats:NegBinomial]{stats::dnbinom()}} is then employed to calculate
the distribution.
}
\usage{
geometric(p, start_at_one = FALSE, eps = 1e-10)
}
\arguments{
\item{p}{Success probability}

\item{start_at_one}{Whether to start the support at 0 or 1 (default: FALSE)}

\item{eps}{How close the distribution is approximated (see details, optional)}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the geometric distribution with the given
parameters.
}
\details{
As the geometric distribution has countably infinite support and this package
only works with discrete distributions with finite support, the resulting
\code{ddf} object approximates the geometric distribution.

For this, the support is cut off at a large enough integer such that the
overall probability is still close to 1.
The cutoff is controlled via the \code{eps} argument which specifies how close the
sum of all probabilities has to be to 1. The default value is 1e-10 since
this is also the minimum accuracy required for creating valid \code{ddf} objects.

Note that both conventions for the geometric distribution (i.e. starting at 0
or 1) are supported by the function via its \code{start_at_one} argument.
}
\examples{
geometric(0.8)
# A more accurate approximation of the same distribution,
# starting at 0 instead of 1
geometric(0.8, TRUE, 1e-15)
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{bin}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
