% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{hypergeometric}
\alias{hypergeometric}
\title{The hypergeometric distribution}
\source{
This function uses \code{\link[stats:Hypergeometric]{stats::dhyper()}} to calculate the distribution
efficiently.
}
\usage{
hypergeometric(N, K, n)
}
\arguments{
\item{N}{Population size}

\item{K}{Number of success states in the population}

\item{n}{Number of draws}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the hypergeometric distribution with the given
parameters.
}
\examples{
# Model the total number of white marbles when
# drawing 5 marbles from an urn containing
# 6 white marbles and 20 marbles in total
hypergeometric(20, 6, 5)
}
