% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-approximated.R
\name{negative_bin}
\alias{negative_bin}
\title{The negative binomial distribution}
\source{
In order to calculate a fitting cutoff, the quantile function
\code{\link[stats:NegBinomial]{stats::qnbinom()}} is used. \code{\link[stats:NegBinomial]{stats::dnbinom()}} is then employed to calculate
the distribution.
}
\usage{
negative_bin(r, p, eps = 1e-10)
}
\arguments{
\item{r}{Number of successes until the experiment is stopped}

\item{p}{Success probability in each experiment}

\item{eps}{How close the distribution is approximated (see details, optional)}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the negative binomial distribution with the given
parameters.
}
\details{
As the negative binomial distribution has countably infinite support and
this package only works with discrete distributions with finite support, the
resulting \code{ddf} object approximates the negative binomial distribution.

For this, the support is cut off at a large enough integer such that the
overall probability is still close to 1.
The cutoff is controlled via the \code{eps} argument which specifies how close the
sum of all probabilities has to be to 1. The default value is 1e-10 since
this is also the minimum accuracy required for creating valid \code{ddf} objects.
}
\examples{
# 'r' does not have to be an integer
negative_bin(2.25, 0.95)
# A more accurate approximation of the same distribution
negative_bin(2.25, 0.95, 1e-12)
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{bin}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_hypergeometric}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
