% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{negative_hypergeometric}
\alias{negative_hypergeometric}
\title{The negative hypergeometric distribution}
\usage{
negative_hypergeometric(N, K, r)
}
\arguments{
\item{N}{A positive integer, the population size.}

\item{K}{A non-negative integer, the number of success states in the
population. Has to be less or equal to \code{N}.}

\item{r}{A non-negative integer, the number of failures until the experiment
is stopped. Has to be less or equal to \code{N}-\code{K}.}
}
\value{
A \code{ddf} distribution as described above.
}
\description{
Create a \code{ddf} object for the negative hypergeometric distribution with the
given parameters.
}
\details{
The negative hypergeometric distribution models the number of successes when
drawing, \emph{without} replacement, elements from a finite population of size
\eqn{N} which contains \eqn{K} success states \emph{until} precisely \eqn{r}
failures have been found.

It has support \eqn{\{0, \dots, K\}} on which its probability mass function
is given by
\deqn{p(k) = \frac{\binom{k+r-1}{k} \binom{N-r-k}{K-k}}{\binom{N}{K}}.}

The beta-binomial distribution provides a generalization of the negative
hypergeometric distribution, see \code{\link[=beta_binomial]{beta_binomial()}}.
}
\examples{
# Model how many blue marbles are drawn from an urn containing
# 20 marbles of which 6 are blue, when one stops as soon
# as one has found 5 non-blue marbles
negative_hypergeometric(20, 6, 5)
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{bin}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{pois}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
