% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{negative_hypergeometric}
\alias{negative_hypergeometric}
\title{The negative hypergeometric distribution}
\usage{
negative_hypergeometric(N, K, r)
}
\arguments{
\item{N}{Population size}

\item{K}{Number of success states in the population}

\item{r}{Number of failures until the experiment is stopped}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the negative hypergeometric distribution with the
given parameters.
}
\examples{
hypergeometric(20, 6, 5)
}
