% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-approximated.R
\name{pois}
\alias{pois}
\title{The Poisson distribution}
\source{
In order to calculate a fitting cutoff, the quantile function
\code{\link[stats:Poisson]{stats::qpois()}} is used. \code{\link[stats:Poisson]{stats::dpois()}} is then employed to calculate
the distribution.
}
\usage{
pois(lambda, eps = 1e-10)
}
\arguments{
\item{lambda}{A positive number, the expected rate of occurrences.}

\item{eps}{A positive number, how close the distribution is approximated.
See ‘Details.’ (optional)}
}
\value{
A \code{ddf} distribution as described above.
}
\description{
Create a \code{ddf} object for the Poisson distribution with the given parameters.
}
\details{
The Poisson distribution models the number of occurrences of an event within
a given time frame. For this, the events have to occur independently, must
not be able to occur at the same instant and have to occur at a constant
average rate, expressed by the parameter \eqn{\lambda}.

It has support \eqn{\mathbb{N}_0} on which its probability mass function is
given by
\deqn{p(k) = e^{-\lambda} \frac{\lambda^k}{k!}.}

As the Poisson distribution has countably infinite support and this package
only works with discrete distributions with finite support, the resulting
\code{ddf} object can only approximate the Poisson distribution.

For this, the support is cut off at a large enough integer such that the
overall probability is still close to 1.
The cutoff is controlled via the \code{eps} argument which specifies how close the
sum of all probabilities has to be to 1. The default value is 1e-10 since
this is also the minimum accuracy required for creating valid \code{ddf} objects.
}
\examples{
pois(0.3)
# A more accurate approximation of the same distribution
pois(0.3, 1e-15)
}
\seealso{
Other distributions: 
\code{\link{benford}()},
\code{\link{bernoulli}()},
\code{\link{beta_binomial}()},
\code{\link{bin}()},
\code{\link{geometric}()},
\code{\link{hypergeometric}()},
\code{\link{negative_bin}()},
\code{\link{negative_hypergeometric}()},
\code{\link{rademacher}()},
\code{\link{unif}()},
\code{\link{zipf}()}
}
\concept{distributions}
