% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-class.R
\name{probs}
\alias{probs}
\alias{probs,ddf-method}
\title{Get the probabilities of a \code{ddf} object}
\usage{
probs(dist)

\S4method{probs}{ddf}(dist)
}
\arguments{
\item{dist}{A \code{ddf} object}
}
\value{
Probabilities on the support as a numerical vector.
}
\description{
Get the probabilities of a \code{ddf} object
}
\examples{
probs(ddf(1:3))
probs(ddf(1:3, c(1/2, 1/4, 1/4)))
}
