% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-class.R
\name{supp}
\alias{supp}
\alias{supp,ddf-method}
\title{The support}
\usage{
supp(dist)

\S4method{supp}{ddf}(dist)
}
\arguments{
\item{dist}{A \code{ddf} object, the distribution.}
}
\value{
Support as a numerical vector.
}
\description{
This function returns the support of a given distribution
as a numerical vector.
}
\details{
The support of a function is defined as the points in the
function's domain where it is non-zero.
}
\examples{
supp(ddf(1:3))
}
