% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{unif}
\alias{unif}
\title{The discrete uniform distribution}
\usage{
unif(n, start_at_one = TRUE)
}
\arguments{
\item{n}{Upper bound of the support}

\item{start_at_one}{Whether to start the support at 0 or 1 (default: TRUE)}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the discrete uniform distribution with the given
parameters.
}
\details{
This function only generates the uniform distribution on support {1, ..., n}
or {0, ..., n}. For more complicated supports, it's expected to be much
easier to simply use \code{\link[=ddf]{ddf()}}, rather than using a custom function like this
one, see also the example below.
}
\examples{
# Probability distribution for modelling a
# six-sided dice (uniform on {1, ..., 6})
unif(6)

# For more complicated supports, use ddf
# Same distribution as above (except for description):
ddf(1:6)
# Not possible using unif:
ddf(seq(3,12,3))
}
