% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{zipf}
\alias{zipf}
\title{The Zipf distribution}
\usage{
zipf(N, s = 1)
}
\arguments{
\item{N}{Total number of elements}

\item{s}{Exponent for the inverse power law (see details, default: 1)}
}
\value{
A \code{ddf} distribution as described above
}
\description{
Create a \code{ddf} object for the Zipf distribution with the given parameters.
}
\examples{
zipf(5)
}
